feat!: standardize namespaces
BREAKING CHANGE: Renames package, solution, project paths, and namespaces from Digitall.APower to Digitall.Plugins.
