+++ Directory.Build.props
@@ -30,7 +30,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <None Include="$(MSBuildThisFileDirectory)Licence.md" Pack="true" PackagePath="/"/>
+    <None Include="$(MSBuildThisFileDirectory)LICENSE" Pack="true" PackagePath="/"/>
     <None Include="README.md" Pack="true" PackagePath="/"/>
   </ItemGroup>
