+++ examples/SamplePlugin/ExecutorSample.cs
@@ -1,7 +1,7 @@
 // Copyright (c) DIGITALL Nature.All rights reserved
 // DIGITALL Nature licenses this file to you under the Microsoft Public License.
 
-using Digitall.APower;
+using Digitall.Plugins;
 
 namespace SamplePlugin
 {
