+++ examples/SamplePlugin/SamplePlugin.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\src\Digitall.APower\Digitall.APower.csproj" />
+    <ProjectReference Include="..\..\src\Digitall.Plugins\Digitall.Plugins.csproj" />
   </ItemGroup>
 
   <ItemGroup>
