--- src/Digitall.APower/ExecutionResult.cs
@@ -1,20 +1,0 @@
-// Copyright (c) DIGITALL Nature. All rights reserved
-// DIGITALL Nature licenses this file to you under the Microsoft Public License.
-
-namespace Digitall.APower
-{
-    public enum ExecutionResult
-    {
-        /// <summary>
-        /// </summary>
-        Ok,
-
-        /// <summary>
-        /// </summary>
-        Failure,
-
-        /// <summary>
-        /// </summary>
-        Skipped
-    }
-}
