+++ src/Digitall.Plugins/Digitall.Plugins.csproj
@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <PackageId>dgt.apower</PackageId>
+    <PackageId>Digitall.Plugins</PackageId>
   </PropertyGroup>
 
   <ItemGroup>
