+++ src/Digitall.Plugins/ExecutionResult.cs
@@ -1,0 +1,19 @@
+// Copyright (c) DIGITALL Nature. All rights reserved
+// DIGITALL Nature licenses this file to you under the Microsoft Public License.
+
+namespace Digitall.Plugins;
+
+public enum ExecutionResult
+{
+    /// <summary>
+    /// </summary>
+    Ok,
+
+    /// <summary>
+    /// </summary>
+    Failure,
+
+    /// <summary>
+    /// </summary>
+    Skipped
+}
