+++ src/Digitall.Plugins/README.md
