+++ src/Digitall.Plugins/Services/ILoggingFacade.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using Microsoft.Xrm.Sdk.PluginTelemetry;
+
+namespace Digitall.Plugins.Services;
+
+public interface ILoggingFacade
+{
+    void Log(LogLevel logLevel, string message, params object[] args);
+    void Log(LogLevel logLevel, Exception exception, string message, params object[] args);
+    void LogTrace(string message, params object[] args);
+    void LogDebug(string message, params object[] args);
+    void LogInformation(string message, params object[] args);
+    void LogWarning(string message, params object[] args);
+    void LogWarning(Exception exception, string message, params object[] args);
+    void LogError(string message, params object[] args);
+    void LogError(Exception exception, string message, params object[] args);
+    void LogCritical(string message, params object[] args);
+    void LogCritical(Exception exception, string message, params object[] args);
+}
